pyproject.toml
[project]
name = "restore-pro"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Fusen", email = "fus3ngames@gmail.com"},
]
dependencies = [
    "PySide6>=6.7.2",
    "requests>=2.32.3",
    "webuiapi>=0.9.13",
    "Pillow>=10.4.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm.scripts]
start = "python src/main.py"


[tool.pdm]
distribution = false
